Format y axis ticks labels to the same number of decimal places (#339)
Pad extra zeros on to numbers so that all y axis labels on a given axis display with the same number of decimal places. This will not automatically apply across panels - e.g. if no decimal places are needed on panel "1" and 2 on panel "2", the two sides will display none and 2 respectively.